fix(Functionality): :bug: Fixed async action of updating player weekly records and assignment of weekly prizes. Added self-healing to player record docs.  Improved logging/toasting.
Needed to await a for loop rather than run async/await in a forEach method to maintain async order.  When player records are not found or extra docs are found, the function self-heals to form one document only per user, per season, per week.  Improved logging in multiple places to provide warnings when issues arise.  May be indications of a systematic issue, not just dev leftovers.